Some debugging while installing v2.2.0

+++ distribute.sh
@@ -4,7 +4,7 @@
 # Source this script
 
 # 1. Setup
-export ENV_NAME=$CONDA_DEFAULT_ENV 
+ENV_NAME=$CONDA_DEFAULT_ENV 
 if [[ "$CONDA_DEFAULT_ENV" == "base" ]]; then
     echo "❌ Environment is 'base', please activate the environment that you would like to distribute" 
     return 1
@@ -12,8 +12,8 @@ else
     echo "✅ Environment is '${ENV_NAME}'" 
 fi
 
-export ENV_DIR="/exp/mu2e/data/users/sgrant/EAF/env"
-export YAML_DIR="${ENV_DIR}/yml"
+ENV_DIR="/exp/mu2e/data/users/sgrant/EAF/env"
+YAML_DIR="${ENV_DIR}/yml"
 
 echo "👋 Distribute '${ENV_NAME}'? [Y/n]:"
 read -r CONTINUE_STR
@@ -27,13 +27,16 @@ source ~/.bash_profile
 mamba activate ${ENV_NAME}
 
 # 2. Create YAML and timestamp
-export THIS_YAML="${ENV_DIR}/yml/${ENV_NAME}.yml"
+THIS_YAML="${ENV_DIR}/yml/${ENV_NAME}.yml"
 if [[ -f "${THIS_YAML}" ]]; then 
     echo "👋 ${THIS_YAML} already exists. Continue? [Y/n]:"
     read -r CONTINUE_STR
     if [[ "$CONTINUE_STR" != "Y" ]]; then
         echo "❌ Exiting..."
         return 1
+    else
+        echo "✅ Removing ${THIS_YAML}..."
+        rm ${THIS_YAML}
     fi
 fi
 
@@ -49,10 +52,19 @@ sed -i 's/- pyutils==\([0-9\.]*\)/- "git+https:\/\/github.com\/Mu2e\/pyutils.git
 mv "tmp" "${THIS_YAML}"
 echo "✅ Written '${THIS_YAML}'"
 
-export TIMESTAMP="${ENV_DIR}/yml/${ENV_NAME}.datetime"
+TIMESTAMP="${ENV_DIR}/yml/${ENV_NAME}.datetime"
 if [[ -f "${TIMESTAMP}" ]]; then 
-    rm "${TIMESTAMP}" 
+    echo "👋 ${TIMESTAMP} already exists. Continue? [Y/n]:"
+    read -r CONTINUE_STR
+    if [[ "$CONTINUE_STR" != "Y" ]]; then
+        echo "❌ Exiting..."
+        return 1
+    else
+        echo "✅ Removing ${TIMESTAMP}..."
+        rm "${TIMESTAMP}" 
+    fi
 fi
+
 date +"%Y-%m-%d_%H-%M-%S" > "${TIMESTAMP}"
 echo "✅ Created timestamp '${TIMESTAMP}'"
 
@@ -65,9 +77,30 @@ if [[ "$CONTINUE_STR" != "Y" ]]; then
 fi
 
 if [[ -d ${ENV_DIR}/${ENV_NAME} ]]; then 
-    echo "❌ Error: '${ENV_DIR}/${ENV_NAME}' already exists!"
-    return 1
-fi   
+    echo "👋 ${ENV_DIR}/${ENV_NAME} already exists. Continue? [Y/n]:"
+    read -r CONTINUE_STR
+    if [[ "$CONTINUE_STR" != "Y" ]]; then
+        echo "❌ Exiting..."
+        return 1
+    else
+        echo "✅ Removing ${ENV_DIR}/${ENV_NAME}..."
+        rm -rf "${ENV_DIR}/${ENV_NAME}"
+    fi
+fi
+
+# Check if tar file already exists and remove it
+TAR_FILE="${ENV_DIR}/tar/${ENV_NAME}.tar.gz"
+if [[ -f "${TAR_FILE}" ]]; then 
+    echo "👋 ${TAR_FILE} already exists. Continue? [Y/n]:"
+    read -r CONTINUE_STR
+    if [[ "$CONTINUE_STR" != "Y" ]]; then
+        echo "❌ Exiting..."
+        return 1
+    else
+        echo "✅ Removing old tar file ${TAR_FILE}..."
+        rm "${TAR_FILE}"
+    fi
+fi
 
 echo "✅ Packing '${ENV_NAME}' into '${ENV_DIR}/tar/${ENV_NAME}.tar.gz'..."
 conda pack -o "${ENV_DIR}/tar/${ENV_NAME}.tar.gz"
